requirements.txt
numpy>=1.24.0
matplotlib>=3.7.0
pyyaml>=6.0

# 强化学习依赖
gymnasium>=0.29.0
stable-baselines3>=2.0.0
torch>=2.0.0
tensorboard>=2.14.0
